rename utils to optional imports
