+++ tests/test_xinfer_top_functions.py
@@ -7,8 +7,6 @@ def test_list_models():
 
 
 def test_list_models_interactive():
-    result = xinfer.list_models(interactive=True)
-    assert result is not None
+    xinfer.list_models(interactive=True)
 
-    result = xinfer.list_models(interactive=True, limit=10)
-    assert len(result) == 10 + 2
+    xinfer.list_models(interactive=False, limit=10)
